Move all play button files under one folder
