+++ packages/apollos-ui-connected/src/MediaControlsConnected/PlayButtonConnected/PlayButtonConnected.js
