+++ packages/apollos-ui-connected/src/MediaControlsConnected/PlayButtonConnected/index.js
@@ -1,0 +1,5 @@
+import PlayButtonConnected from './PlayButtonConnected';
+
+export PlayButton from './PlayButton';
+
+export default PlayButtonConnected;
